+++ projects/github-actions/repo-gardening/src/utils/slack/send-slack-message.js
@@ -1,5 +1,5 @@
 const { getInput, setFailed } = require( '@actions/core' );
-const fetch = require( 'node-fetch' );
+const { WebClient, ErrorCode } = require( '@slack/web-api' );
 
 /* global WebhookPayloadPullRequest, WebhookPayloadIssue */
 
@@ -18,6 +18,8 @@ async function sendSlackMessage( message, channel, payload, customMessageFormat 
 		setFailed( 'triage-issues: Input slack_token is required but missing. Aborting.' );
 		return;
 	}
+
+	const slackApi = new WebClient( token );
 
 	let slackMessage = '';
 
@@ -77,18 +79,71 @@ async function sendSlackMessage( message, channel, payload, customMessageFormat 
 		};
 	}
 
-	const slackRequest = await fetch( 'https://slack.com/api/chat.postMessage', {
-		method: 'POST',
-		body: JSON.stringify( slackMessage ),
-		headers: {
-			'Content-Type': 'application/json; charset=utf-8',
-			'Content-Length': slackMessage.length,
-			Authorization: `Bearer ${ token }`,
-			Accept: 'application/json',
-		},
-	} );
+	try {
+		const slackRequest = await slackApi.chat.postMessage( slackMessage );
+		return !! slackRequest.ok;
+	} catch ( error ) {
+		// The request failed.
+		// At this point, we want to log specific types of errors (let's avoid noise by logging temporary errors for example).
+		if ( ! error.code === ErrorCode.PlatformError ) {
+			return false;
+		}
 
-	return !! slackRequest.ok;
+		// See the list of error messages here: https://api.slack.com/methods/chat.postMessage#errors
+		const errorMessage = error?.data?.error ?? 'Unknown error';
+
+		// Let's send a direct message to @jeherve about it, so we can investigate.
+		// For folks outside of Automattic, let's use the Quality team channel.
+		const {
+			repository: { owner },
+		} = payload;
+		const { html_url, title } = payload?.pull_request ?? payload.issue;
+
+		const reportingChannel =
+			owner === 'automattic' ? 'D1KN8VCCA' : getInput( 'slack_quality_channel' );
+		if ( ! reportingChannel ) {
+			return false;
+		}
+		const reportMessage = {
+			channel: reportingChannel,
+			blocks: [
+				{
+					type: 'section',
+					text: {
+						type: 'mrkdwn',
+						text: `We attempted to send a Slack message to ${ reportingChannel } about the issue below, but received the following error: \`${ errorMessage }\``,
+					},
+				},
+				{
+					type: 'divider',
+				},
+				{
+					type: 'section',
+					text: {
+						type: 'mrkdwn',
+						text: `<${ html_url }|${ title }>`,
+					},
+					accessory: {
+						type: 'button',
+						text: {
+							type: 'plain_text',
+							text: 'Review',
+							emoji: true,
+						},
+						value: 'click_review',
+						url: `${ html_url }`,
+						action_id: 'button-action',
+					},
+				},
+			],
+			text: `Error sending message to Slack: -- <${ html_url }|${ title }>`, // Fallback text for display in notifications.
+			mrkdwn: true, // Formatting of the fallback text.
+			unfurl_links: false,
+			unfurl_media: false,
+		};
+		const reportMessageSlackResponse = await slackApi.chat.postMessage( reportMessage );
+		return !! reportMessageSlackResponse.ok;
+	}
 }
 
 module.exports = sendSlackMessage;