+++ projects/packages/my-jetpack/_inc/admin.jsx
@@ -17,7 +17,6 @@ import {
 	BackupInterstitial,
 	BoostInterstitial,
 	CRMInterstitial,
-	CreatorInterstitial,
 	ExtrasInterstitial,
 	JetpackAiInterstitial,
 	ProtectInterstitial,
@@ -26,6 +25,9 @@ import {
 	SearchInterstitial,
 	VideoPressInterstitial,
 	StatsInterstitial,
+	SecurityInterstitial,
+	GrowthInterstitial,
+	CompleteInterstitial,
 } from './components/product-interstitial';
 import JetpackAiProductPage from './components/product-interstitial/jetpack-ai/product-page';
 import RedeemTokenScreen from './components/redeem-token-screen';
@@ -70,7 +72,6 @@ const MyJetpack = () => {
 								<Route path={ MyJetpackRoutes.AddBackup } element={ <BackupInterstitial /> } />
 								<Route path={ MyJetpackRoutes.AddBoost } element={ <BoostInterstitial /> } />
 								<Route path={ MyJetpackRoutes.AddCRM } element={ <CRMInterstitial /> } />
-								<Route path={ MyJetpackRoutes.AddCreator } element={ <CreatorInterstitial /> } />
 								<Route
 									path={ MyJetpackRoutes.AddJetpackAI }
 									element={ <JetpackAiInterstitial /> }
@@ -89,8 +90,11 @@ const MyJetpack = () => {
 									<Route path={ MyJetpackRoutes.AddLicense } element={ <AddLicenseScreen /> } />
 								) }
 								<Route path={ MyJetpackRoutes.RedeemToken } element={ <RedeemTokenScreen /> } />
-								<Route path="/redeem-token" element={ <RedeemTokenScreen /> } />
-								<Route path="/jetpack-ai" element={ <JetpackAiProductPage /> } />
+								<Route path={ MyJetpackRoutes.RedeemToken } element={ <RedeemTokenScreen /> } />
+								<Route path={ MyJetpackRoutes.JetpackAi } element={ <JetpackAiProductPage /> } />
+								<Route path={ MyJetpackRoutes.AddSecurity } element={ <SecurityInterstitial /> } />
+								<Route path={ MyJetpackRoutes.AddGrowth } element={ <GrowthInterstitial /> } />
+								<Route path={ MyJetpackRoutes.AddComplete } element={ <CompleteInterstitial /> } />
 							</Routes>
 						</HashRouter>
 					</QueryClientProvider>