+++ projects/packages/my-jetpack/global.d.ts
@@ -39,7 +39,9 @@ type JetpackModule =
 	| 'security'
 	| 'protect'
 	| 'videopress'
-	| 'stats';
+	| 'stats'
+	| 'growth'
+	| 'complete';
 
 type ThreatItem = {
 	// Protect API properties (free plan)
@@ -155,6 +157,8 @@ interface Window {
 					plugin_slug: string;
 					post_activation_url: string;
 					post_checkout_url?: string;
+					manage_paid_plan_purchase_url?: string;
+					renew_paid_plan_purchase_url?: string;
 					pricing_for_ui?: {
 						available: boolean;
 						wpcom_product_slug: string;
@@ -172,6 +176,11 @@ interface Window {
 							should_prorate_when_offer_ends: boolean;
 							transition_after_renewal_count: number;
 							usage_limit?: number;
+							reason?: {
+								errors: {
+									introductoryOfferRemovedSubscriptionFound: string[];
+								};
+							};
 						};
 						tiers?: {
 							[ key: string ]: {
@@ -186,6 +195,11 @@ interface Window {
 									shouldProrateWhenOfferEnds: boolean;
 									transitionAfterRenewalCount: number;
 									usageLimit?: number;
+									reason?: {
+										errors: {
+											introductoryOfferRemovedSubscriptionFound: string[];
+										};
+									};
 								};
 								isIntroductoryOffer: boolean;
 								productTerm: string;