+++ projects/js-packages/api/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [0.17.19] - 2024-11-25
+### Changed
+- Updated package dependencies. [#40288]
+
 ## [0.17.18] - 2024-11-14
 ### Changed
 - Update dependencies.
@@ -92,8 +96,7 @@
 
 ## [0.16.7] - 2023-12-03
 ### Changed
-- Updated package dependencies. [#34411]
-- Updated package dependencies. [#34427]
+- Updated package dependencies. [#34411] [#34427]
 
 ## [0.16.6] - 2023-11-20
 
@@ -374,6 +377,7 @@
 - Add the API methods left behind by the previous PR.
 - Initial release of jetpack-api package
 
+[0.17.19]: https://github.com/Automattic/jetpack-api/compare/v0.17.18...v0.17.19
 [0.17.18]: https://github.com/Automattic/jetpack-api/compare/v0.17.17...v0.17.18
 [0.17.17]: https://github.com/Automattic/jetpack-api/compare/v0.17.16...v0.17.17
 [0.17.16]: https://github.com/Automattic/jetpack-api/compare/v0.17.15...v0.17.16