+++ projects/packages/plugin-deactivation/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-11-28
+### Removed
+- General: Update minimum PHP version to 7.2. [#40147]
+
 ## [0.2.4] - 2024-11-04
 ### Added
 - Enable test coverage. [#39961]
@@ -15,8 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.2] - 2024-08-29
 ### Changed
-- Updated package dependencies. [#39004]
-- Updated package dependencies. [#39111]
+- Updated package dependencies. [#39004] [#39111]
 
 ## [0.2.1] - 2024-03-15
 ### Changed
@@ -60,6 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added package to intercept plugin deactivation [#27081]
 
+[0.3.0]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.2.4...v0.3.0
 [0.2.4]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.2.1...v0.2.2